Mini Project 2
This application is designed to perform CRUD Operations.

In computer programming, create, read, update and delete (as an acronym
CRUD or possibly a backronym) (Sometimes called SCRUD with an "S" for
Search) are the four basic functions of persistent storage. Sometimes
CRUD is expanded with the words retrieve instead of read, modify instead
of update, or destroy instead of delete. It is also sometimes used to
describe user interface conventions that facilitate viewing, searching,
and changing information; often using computer-based forms and reports.

Mp2-Readme.docx is the ReadMe file for the Project